requirements-dev.txt
django-debug-toolbar==5.0.1

ruff==0.9.3
mypy==1.14.1
isort==6.0.0

# Types
django-stubs==5.1.2
